Renamed internal version of population and stochastic files to have a trailing "_internal"
