+++ Source/Factory/genetic_factory.cpp
@@ -1,7 +1,7 @@
 #include <vector>
 
 #include "../API/Factory/genetic_factory.h"
-//#include "../population.h"
+//#include "../population_internal.h"
 
 population* genetic_factory::create_population(population_metadata* population_metadata)
 {
