+++ Source/stochastic_internal.h
